requirements.txt
torch==0.1.1.0
bidict>=0.17.2
biopython>=1.71
Deprecated>=1.2.3
numpy>=1.14.3
pandas>=0.23.0
py>=1.5.3
pytorch-ignite>=0.1.0
SQLAlchemy>=1.3.1
SQLAlchemy-Utils==0.33.5
torchfile>=0.1.0
